// src/lib/api/generated/model/currentUser.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Django Continuous Integration API
 * API documentation for the Django Continuous Integration project with tenant-aware access rules.
 * OpenAPI spec version: 1.0.0
 */
import type { TenantMembership } from './tenantMembership';

export interface CurrentUser {
  readonly id: string;
  /** @maxLength 254 */
  email: string;
  /** @maxLength 150 */
  first_name?: string;
  /** @maxLength 150 */
  last_name?: string;
  is_active?: boolean;
  readonly current_tenant: string;
  readonly memberships: readonly TenantMembership[];
}