// src/lib/api/generated/model/tokenPairResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Django Continuous Integration API
 * API documentation for the Django Continuous Integration project with tenant-aware access rules.
 * OpenAPI spec version: 1.0.0
 */

export interface TokenPairResponse {
  refresh: string;
  access: string;
}